+++ Project5/src/main/scala/Objects/SecureMessage.scala
@@ -1,7 +1,7 @@
 package Objects
 
 case class SecureMessage(
-                          from: Long,
+                          from: Int,
                           message: Array[Byte],
                           signature: Array[Byte],
                           encryptedKey: Array[Byte]
